audit

+++ controller/client/client.controller.js
@@ -1,9 +1,10 @@
 import db from '../../models/index.js';
-const { Clients, Contacts } = db;
+const { Clients, Contacts, Audit } = db;
 
 export const createClient = async (req, res) => {
    try {
     const client = await Clients.create(req.body);
+    await Audit.create({ userId: req.body.userId, action: 'create', tableName: 'clients', newData: client.get() });
     res.status(201).json(client);
    } catch (error) {
     res.status(500).json({ message: error.message });
@@ -32,6 +33,7 @@ export const updateClient = async (req, res) => {
     try {
         const client = await Clients.findByPk(req.params.id);
         client.update(req.body);
+        await Audit.create({ userId: req.body.userId, action: 'update', tableName: 'clients', oldData: client.get(), newData: req.body });
         res.status(200).json(client);
     } catch (error) {
         res.status(500).json({ message: error.message });
@@ -40,7 +42,12 @@ export const updateClient = async (req, res) => {
 
 export const deleteClient = async (req, res) => {
     try {
+        const client = await Clients.findByPk(req.params.id);
+        if (!client) {
+            return res.status(404).json({ message: 'Client not found' });
+        }
         await Clients.destroy({ where: { id: req.params.id } });
+        await Audit.create({ userId: req.body.userId, action: 'delete', tableName: 'clients', oldData: client.get() });
         res.status(200).json({ message: "Client deleted successfully" });
     } catch (error) {
         res.status(500).json({ message: error.message });
@@ -68,6 +75,7 @@ export const statusUpdate = async (req, res) => {
     try {
         const client = await Clients.findByPk(req.params.id);
         client.update({ accountStatus: req.body.status });
+        await Audit.create({ userId: req.body.userId, action: 'update', tableName: 'clients', oldData: client.get(), newData: req.body });
         res.status(200).json(client);
     } catch (error) {
         res.status(500).json({ message: error.message });
