+++ controller/user.module/role.controller.js
@@ -1,11 +1,12 @@
 import db from '../../models/index.js';
-const { Role } = db;
+const { Role, Audit } = db;
 
 export const createRole = async (req, res) => {
     try {
         const { name } = req.body;
         console.log(name);
         const role = await Role.create({ name });
+        await Audit.create({ userId: req.body.userId, action: 'create', tableName: 'roles', newData: role.get() });
         res.status(201).json(role);
     } catch (error) {
         console.log(error);
@@ -35,6 +36,7 @@ export const updateRole = async (req, res) => {
         }
         role.name = name;
         await role.save();
+        await Audit.create({ userId: req.body.userId, action: 'update', tableName: 'roles', oldData: role.get(), newData: req.body });
         res.status(200).json(role);
     } catch (error) {
         res.status(500).json({ error: error.message });
@@ -45,7 +47,12 @@ export const updateRole = async (req, res) => {
 export const deleteRole = async (req, res) => {
     try {
         const { id } = req.params;
+        const role = await Role.findByPk(id);
+        if (!role) {
+            return res.status(404).json({ error: 'Role not found' });
+        }
         await Role.destroy({ where: { id } });
+        await Audit.create({ userId: req.body.userId, action: 'delete', tableName: 'roles', oldData: role.get() });
         res.status(200).json({ message: 'Role deleted successfully' });
     } catch (error) {
         res.status(500).json({ error: error.message });
